v0.4.9 [General] - Fixed images not showing properly on edit - Fixed images causing errors when the files does not exist - Temporarily disabled 2FA feature

+++ app/Livewire/Coordinator/Submissions.php
@@ -730,7 +730,7 @@ class Submissions extends Component
                 ->get();
 
             foreach ($credentials as $credential) {
-                if (isset($credential->image_file_path)) {
+                if (isset($credential->image_file_path) && Storage::exists($credential->image_file_path)) {
                     Storage::delete($credential->image_file_path);
                 }
